+++ Tests/ActionTests/ActionTests.swift
@@ -16,19 +16,21 @@ class ActionTests: QuickSpec {
 		}
 
         describe("completable action") {
+            var inputs: TestableObserver<String>!
             var action: CompletableAction<String>!
             beforeEach {
-                let work: TestableObservable<Never> = scheduler.createColdObservable([.completed(0)])
-                action = CompletableAction {_ in work.asCompletable() }
+                inputs = scheduler.createObserver(String.self)
+                action = CompletableAction { input -> Completable in
+                    inputs.onNext(input)
+                    return Observable<Never>.empty().asCompletable()
+                }
                 scheduler.scheduleAt(10) { action.inputs.onNext("a") }
                 scheduler.scheduleAt(20) { action.inputs.onNext("b") }
             }
             afterEach {
                 action = nil
             }
-            it("inputs subject receives generated inputs") {
-                let inputs = scheduler.createObserver(String.self)
-                action.inputs.bind(to: inputs).disposed(by: disposeBag)
+            it("receives generated inputs") {
                 scheduler.start()
                 XCTAssertEqual(inputs.events, [
                     next(10, "a"),
@@ -47,6 +49,72 @@ class ActionTests: QuickSpec {
                 scheduler.start()
                 XCTAssert(completions.events.contains { $0.time == 10})
                 XCTAssert(completions.events.contains { $0.time == 20})
+            }
+        }
+
+        describe("Input observer behavior") {
+            var action: Action<String, String>!
+            var inputs: TestableObserver<String>!
+            var executions: TestableObserver<Observable<String>>!
+            beforeEach {
+                inputs = scheduler.createObserver(String.self)
+                action = Action {
+                    inputs.onNext($0)
+                    return Observable.just($0)
+                }
+                executions = scheduler.createObserver(Observable<String>.self)
+                action.executionObservables.bind(to: executions).disposed(by: disposeBag)
+            }
+            afterEach {
+                action = nil
+                inputs = nil
+                executions = nil
+            }
+
+            it("execute on .next") {
+                scheduler.scheduleAt(10) { action.inputs.onNext("a") }
+                scheduler.start()
+                XCTAssertEqual(inputs.events, [next(10, "a")])
+                XCTAssertEqual(executions.events.filter { !$0.value.isStopEvent }.count, 1)
+            }
+            it("ignore .error events") {
+                scheduler.scheduleAt(10) { action.inputs.onError(TestError) }
+                scheduler.start()
+                XCTAssertEqual(inputs.events, [])
+                XCTAssertEqual(executions.events.filter { !$0.value.isStopEvent }.count, 0)
+            }
+            it("ignore .completed events") {
+                scheduler.scheduleAt(10) { action.inputs.onCompleted() }
+                scheduler.start()
+                XCTAssertEqual(inputs.events, [])
+                XCTAssertEqual(executions.events.filter { !$0.value.isStopEvent }.count, 0)
+            }
+            it("accept multiple .next events") {
+                scheduler.scheduleAt(10) { action.inputs.onNext("a") }
+                scheduler.scheduleAt(20) { action.inputs.onNext("b") }
+                scheduler.start()
+                XCTAssertEqual(inputs.events, [
+                    next(10, "a"),
+                    next(20, "b"),
+                    ])
+                XCTAssertEqual(executions.events.filter { !$0.value.isStopEvent }.count, 2)
+                XCTAssertEqual(executions.events.filter { $0.value.isStopEvent }.count, 0)
+            }
+            it("not terminate after .error event") {
+                scheduler.scheduleAt(10) { action.inputs.onError(TestError) }
+                scheduler.scheduleAt(20) { action.inputs.onNext("b") }
+                scheduler.start()
+                XCTAssertEqual(inputs.events, [next(20, "b")])
+                XCTAssertEqual(executions.events.filter { !$0.value.isStopEvent }.count, 1)
+                XCTAssertEqual(executions.events.filter { $0.value.isStopEvent }.count, 0)
+            }
+            it("not terminate after .completed event") {
+                scheduler.scheduleAt(10) { action.inputs.onCompleted() }
+                scheduler.scheduleAt(20) { action.inputs.onNext("b") }
+                scheduler.start()
+                XCTAssertEqual(inputs.events, [next(20, "b")])
+                XCTAssertEqual(executions.events.filter { !$0.value.isStopEvent }.count, 1)
+                XCTAssertEqual(executions.events.filter { $0.value.isStopEvent }.count, 0)
             }
         }
         
@@ -68,12 +136,14 @@ class ActionTests: QuickSpec {
                 executionObservables = scheduler.createObserver(Observable<String>.self)
                 underlyingError = scheduler.createObserver(Error.self)
             }
-			
-			func bindAction(action: Action<String, String>) {
-				action.inputs
-					.bind(to: inputs)
-					.disposed(by: disposeBag)
-				
+
+            func buildAction(enabledIf: Observable<Bool> = Observable.just(true),
+                             factory: @escaping (String) -> Observable<String>) -> Action<String, String> {
+                let action = Action<String, String>(enabledIf: enabledIf) {
+                    inputs.onNext($0)
+                    return factory($0)
+                }
+
 				action.elements
 					.bind(to: elements)
 					.disposed(by: disposeBag)
@@ -98,19 +168,20 @@ class ActionTests: QuickSpec {
                     .bind(to: underlyingError)
                     .disposed(by: disposeBag)
 				
-                // Dummy subscription for multiple subcription tests
-                action.inputs.subscribe().disposed(by: disposeBag)
+				// Dummy subscription for multiple subcription tests
                 action.elements.subscribe().disposed(by: disposeBag)
                 action.errors.subscribe().disposed(by: disposeBag)
                 action.enabled.subscribe().disposed(by: disposeBag)
                 action.executing.subscribe().disposed(by: disposeBag)
                 action.executionObservables.subscribe().disposed(by: disposeBag)
                 action.underlyingError.subscribe().disposed(by: disposeBag)
+
+                return action
 			}
 			
 			describe("single element action") {
 				sharedExamples("send elements to elements observable") {
-					it("inputs subject receives generated inputs") {
+					it("work factory receives inputs") {
 						XCTAssertEqual(inputs.events, [
 							next(10, "a"),
 							next(20, "b"),
@@ -156,8 +227,7 @@ class ActionTests: QuickSpec {
 				var action: Action<String, String>!
 				
 				beforeEach {
-					action = Action { Observable.just($0) }
-					bindAction(action: action)
+					action = buildAction { Observable.just($0) }
 				}
 				
 				context("trigger via inputs subject") {
@@ -183,7 +253,7 @@ class ActionTests: QuickSpec {
 			
 			describe("multiple element action") {
 				sharedExamples("send array elements to elements observable") {
-					it("inputs subject receives generated inputs") {
+					it("work factory receives inputs") {
 						XCTAssertEqual(inputs.events, [
 							next(10, "a"),
 							next(20, "b"),
@@ -233,7 +303,7 @@ class ActionTests: QuickSpec {
 				var action: Action<String, String>!
 				
 				beforeEach {
-					action = Action { input in
+					action = buildAction { input in
 						// "a" -> ["a", "b", "c"]
 						let baseValue = UnicodeScalar(input)!.value
 						let strings = (baseValue..<(baseValue + 3))
@@ -242,8 +312,6 @@ class ActionTests: QuickSpec {
 						
 						return Observable.from(strings)
 					}
-					
-					bindAction(action: action)
 				}
 				
 				context("trigger via inputs subject") {
@@ -269,7 +337,7 @@ class ActionTests: QuickSpec {
 			
 			describe("error action") {
 				sharedExamples("send errors to errors observable") {
-					it("inputs subject receives generated inputs") {
+					it("work factory receives inputs") {
 						XCTAssertEqual(inputs.events, [
 							next(10, "a"),
 							next(20, "b"),
@@ -326,8 +394,7 @@ class ActionTests: QuickSpec {
 				var action: Action<String, String>!
 				
 				beforeEach {
-					action = Action { _ in Observable.error(TestError) }
-					bindAction(action: action)
+					action = buildAction { _ in Observable.error(TestError) }
 				}
 				
 				context("trigger via inputs subject") {
@@ -353,11 +420,8 @@ class ActionTests: QuickSpec {
 			
 			describe("disabled action") {
 				sharedExamples("send notEnabled errors to errors observable") {
-					it("inputs subject receives generated inputs") {
-						XCTAssertEqual(inputs.events, [
-							next(10, "a"),
-							next(20, "b"),
-							])
+					it("work factory receives nothing") {
+						XCTAssertEqual(inputs.events, [])
 					}
 					
 					it("elements observable receives nothing") {
@@ -395,8 +459,7 @@ class ActionTests: QuickSpec {
 				var action: Action<String, String>!
 				
 				beforeEach {
-					action = Action(enabledIf: Observable.just(false)) { Observable.just($0) }
-					bindAction(action: action)
+					action = buildAction(enabledIf: Observable.just(false)) { Observable.just($0) }
 				}
 				
 				context("trigger via inputs subject") {